Merge LibvirtOpenVswitchVirtualPortDriver into LibvirtOpenVswitchDriver
The LibvirtOpenVswitchVirtualPortDriver VIF driver and the
LibvirtOpenVswitchDriver VIF driver both support OpenVswitch
based guest configs. The only difference is that the former
uses libvirt's modern <network type='bridge'> with vport
profile specified, while the latter uses <network type='ethernet'>
and does all openvswitch setup itself. The reason for the latter
is to support running against libvirt versions older than 0.9.11

Rather than require the host admin to figure this out themselves,
merge the two classes together and automatically choose the right
impl based on the libvirt version we're running against

The LibvirtOpenVswitchVirtualPortDriver class is kept around as
a trivial subclass of LibvirtOpenVswitchDriver for backwards
compatibility. It will be removed in the Hxxx release.

Change-Id: I1fb357ef9de35168d936e3c6dcbf0d4efc3692f6
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>