Merge LibvirtHybridOVSBridgeDriver into LibvirtOpenVswitchDriver
The LibvirtHybridOVSBridgeDriver VIF driver and the
LibvirtOpenVswitchDriver VIF driver both support OpenVswitch
based guest configs. The difference is that the hybrid driver
does not connect the VM directly to the OVS bridge. Instead it
connects the guest to a private Linux bridge, connected to the
OVS bridge via a veth pair. This allows iptables firewall rules
to be applied to the guest VIF. The hybrid driver should always
be used unless the admin has disabled Nova firewalls using the
NoopFirewallDriver impl.

Rather than require the host admin to figure this out themselves,
merge the two classes together and automatically choose the right
impl based on whether NoopFirewallDriver is in use.

When Quantum gains support for firewalling, it will need to be
able to inform Nova whether it has got a filtering applied to
the VIFs, and thus let Nova intelligently skip its own attempt
at filtering.

The LibvirtHybridOVSBridgeDriver class is kept around as
a trivial subclass of LibvirtOpenVswitchDriver for backwards
compatibility. It will be removed in the Hxxx release.

Change-Id: I1fb357ef9de35168d936e3c6dcbf0d4efc3692f6
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>