Gut the QuantumLinuxBridgeVIFDriver implementation
The QuantumLinuxBridgeVIFDriver no longer has any quantum
specific code in it, so the original LibvirtBridgeDriver
can be used with Quantum directly. This has the advantage
that libvirt firewall filter parameters are correctly
configured.

For backwards compatibility make the QuantumLinuxBridgeVIFDriver
class be a trivial subclass of LibvirtBridgeDriver. Log a warning
that it will be removed in the Hxxxx release series.

Change-Id: I19664f20e5137f495513999214f66279292bda58
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>