Introduce basic data model for guru meditation reports
This introduces a new namespace openstack.common.gurumed.
This package will contain modules related to generation
of guru meditation reports, which provide details on all
system state. This commit includs the first part of the
impl, which is the 'model' module containing a number
of classes for the core data models involved.

Blueprint guru-meditation-reports
Change-Id: I5d81cbe73f5646c4932edd1cf1ef6fd8aedda428
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>