Trigger a guru meditation report on SIGUSR1 for all binaries
Make all binaries activate support for automatically dumping
a guru meditation report on SIGUSR1. This is useful for
debugging of production systems where it is not practical to
have the service interruption inherant in attaching a debugger
to the process. Upon receipt of SIGUSR1, a dump will be sent
to the daemon's stderr stream.

Blueprint guru-meditation-reports
Change-Id: Ie4976d434c8d6c27c70a56c6466e47f1106a6678
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>