libvirt: fix overly strict CPU model comparison in live migration
The current libvirt driver migration code checks whether the
dest host can support the guest, by comparing the source host
CPU model. This is overly strict, as the guest may well be
running with a more restrictive CPU model that hides many
host features.  The correct approach is to compare against
the guest's configured CPU model.

To achieve this is it neccessary to introduce a new virt
driver method to query a guest's CPU model. For this API
we define a simple class to represent the CPU model that
is currently able to support the three different ways that
libvirt can represent a CPU model. In the future other virt
drivers can use this same class, or extend it to add new
ways to represent the CPU model if needed.

To allow guests to migrate from new hosts to old hosts,
the new host must still report the 'cpu_info' data against
the host service record. To allow guests to migrate from
old hosts to new hosts, the new host must fallback to using
the 'cpu_info' data from the src host service record, if no
guest CPU was given.

Co-authored: Daniel Berrange <berrange@redhat.com>
Closes-Bug: 1082414
Change-Id: I65f505fec64c65d2641d2bfd940cde44bcd83a78