virt: introduce VirtInstanceInfo object
Formalize the reporting of supported instance information in the
virt drivers by introducing a VirtInstanceInfo object. The
'supported_instances' JSON string in the dict returned
by get_available_resource is now a list of VirtInstanceInfo
objects. The resource tracker then converts this to the
original JSON format when it needs to.

Change-Id: Ia738cf378d0c2e8420d2cd0ec0ddf82d5f1ae79c