virt: introduce VirtPCIDeviceInfo & VirtPCIAddressInfo classes
Extend nova.virt.hardware to provide the VirtPCIDeviceInfo and
VirtPCIAddressInfo classes. The VirtPCIDeviceInfo class provides
information about a single PCI device present on the host. It
references a VirtPCIAddressInfo class instance to provide the
address of the device. If the device is a virtual function,
it will also referene a second VirtPCIAddressInfo instance to
provide the address of the physical function.

Change-Id: I27ed23c6a7671b9bf3257f2af4437974daa0b11a